Squashed 'client-rn/' changes from 7daf6db..e231567
e231567 Publish
3df9078 [client-ts] Fix valid signature to handle new multi-sig implementation
77bc5f3 [client-ts] Fix closing local databases not waiting for remote sync to complete
2a62573 [client-ts] Fix database closing when sending inbox messages
1cc3510 Publish
3ef67f1 Update CHANGELOG with profile website verification details
240d8df Update data signing to support multiple sig schemes (#368)
55a84c0 Support verification of profile domains (#367)
86a0772 Revert "Publish"
489ce0c Fix version numbers to reference rc1
aa9f82a Publish
b68408b Update CHANGELOG for release
6a523eb Fix typescript build errors (#363)
24c8ec6 Improve CHANGELOG messages in preparation for release
d6de107 Export migration utils
869bbd6 Feature/353 support replicating contexts (#354)
92e7f3a Close outbox database connection after sending an inbox message
1dc04c2 Publish
e75ef51 Publish
a09b9c4 Publish
2f7d175 Publish
6fc706a Update CHANGELOG.md files for 2.3.5 release
47ade07 [did-client] Add unit test for easily testing a group of storage nodes
80ba807 Publish
b9f828c Publish
de6d0b9 Fix: Getting data by URI didn't support `devnet`
2bea2fc Publish
d1800f0 [vda-common] Update contract name (#318)

git-subtree-dir: client-rn
git-subtree-split: e231567faea2f601d530f08edf0af3e67debad18

+++ src/context/engines/verida/database/db-encrypted.ts
@@ -115,23 +115,25 @@ class EncryptedDatabase extends BaseDb {
      * If there is data in this database, it ensures the current encryption key
      * can decrypt the data.
      */
-    try {
-      await this.getMany();
-      //console.log(`Db.init-5(${databaseName}): ${(new Date()).getTime()-now}`)
-    } catch (err: any) {
-      // This error message is thrown by the underlying decrypt library if the
-      // data can't be decrypted
-      if (
-        err.message == `Unsupported state or unable to authenticate data` ||
-        err.message == "Could not decrypt!"
-      ) {
-        // Clear the instantiated PouchDb instances and throw a more useful exception
-        await this.close()
-        throw new Error(`Invalid encryption key supplied`);
-      }
+    if (this.config.verifyEncryptionKey) {
+      try {
+        await this.getMany();
+        //console.log(`Db.init-5(${databaseName}): ${(new Date()).getTime()-now}`)
+      } catch (err: any) {
+        // This error message is thrown by the underlying decrypt library if the
+        // data can't be decrypted
+        if (
+          err.message == `Unsupported state or unable to authenticate data` ||
+          err.message == "Could not decrypt!"
+        ) {
+          // Clear the instantiated PouchDb instances and throw a more useful exception
+          await this.close()
+          throw new Error(`Invalid encryption key supplied`);
+        }
 
-      // Unknown error, rethrow
-      throw err;
+        // Unknown error, rethrow
+        throw err;
+      }
     }
 
     //console.log(`Db.init-Final(${databaseName}): ${(new Date()).getTime()-now}`)
@@ -228,17 +230,33 @@ class EncryptedDatabase extends BaseDb {
   public async close(options: DatabaseCloseOptions = {
     clearLocal: false
   }) {
+    if (this._sync === null) {
+      // No sync object indicates this database is closed
+      return
+    }
+
     if (options.clearLocal) {
       await this.destroy({
         localOnly: true
       })
 
+      if (this._sync) {
+        this._sync.cancel();
+      }
+  
+      this._sync = null;
+      this._syncError = null;
+
+      // Return, because destroy will close all database connections
       return
     }
 
     if (this._sync) {
       this._sync.cancel();
     }
+
+    this._sync = null;
+    this._syncError = null;
 
     try {
       await this._localDbEncrypted.close();
@@ -252,8 +270,6 @@ class EncryptedDatabase extends BaseDb {
       // may already be closed
     }
 
-    this._sync = null;
-    this._syncError = null;
     await this.engine.closeDatabase(this.did, this.databaseName)
     this.emit('closed', this.databaseName)
   }
@@ -290,20 +306,25 @@ class EncryptedDatabase extends BaseDb {
     localOnly: false
   }): Promise<void> {
     try {
-      // Destory the local pouch database (this deletes this._local and this._localDbEncrypted as they share the same underlying data source)
+      // Destroy the local pouch database (this deletes this._local and this._localDbEncrypted as they share the same underlying data source)
       await this._localDbEncrypted.destroy()
+    } catch (err) {
+      // do nothing, database is likely already destroyed
+    }
 
+    try {
       if (!options.localOnly) {
         // Only delete remote database if required
         await this.engine.deleteDatabase(this.databaseName)
       }
-     
+
       await this.close({
         clearLocal: false
       })
     } catch (err) {
-      console.error(err)
+      console.log(err)
     }
+    
   }
 
   public async updateUsers(
